Move the getBearerToken to a new file.
